fix: glyph decoder sign-extended high CP437 codes, corrupting the title art
The LITERAL_CHAR decoder read the embedded glyph's low byte as a signed
char, so any glyph id whose low byte is >= 128 sign-extended into an
out-of-range id and drew as garbage. Cast the low byte with (uint8) at
both decode sites in src/TextTerm.cpp. Latent upstream defect: char is
signed on Win32/MSVC too, so it corrupts the same ids there. Marked
upstream: and ledgered in docs/REPORTING-GATE.md.

inc-ttte

+++ src/TextTerm.cpp
@@ -232,7 +232,10 @@ void TextTerm::Write(const char*str)
 			ch++;
 			gid = ~(*ch) << 8;
 			ch++;
-			gid |= *ch;
+			// upstream: Win32/MSVC's signed char sign-extends this low byte too.
+			// Traced: 4072 == 0xFE8 from rune 0xE8 (232).
+			// inc-ttte; not sent.
+			gid |= (uint8)*ch;
 			APutChar(cx, cy, GLYPH_VALUE(gid, attr));
 		} else if (*ch == WRAP_BREAK) 
           cWrap = cx+1;
@@ -836,7 +839,10 @@ void TextTerm::SWrite(const char *text, int16 wn)
 			ch++;
 			gid = ~(*ch) << 8;
 			ch++;
-			gid |= *ch;
+			// upstream: Win32/MSVC's signed char sign-extends this low byte too.
+			// Traced: 4072 == 0xFE8 from rune 0xE8 (232).
+			// inc-ttte; not sent.
+			gid |= (uint8)*ch;
 			SPutChar(scx, scy, GLYPH_VALUE(gid, attr));
 		}
         else if (*ch == WRAP_BREAK)
